- Due to the refactoring that has been carried out on the "Big VO" for requisition the variable type that we use for the number of fields has been changed to int so consequent changes to all models APC-1833
git-svn-id: svn://srl009.adg.internal/jlib/trunk@738 bee1aa47-a06f-436c-a2f2-6a1efea3c100

+++ src/main/java/com/areen/jlib/model/SimpleObject.java
@@ -19,7 +19,7 @@ import java.io.Serializable;
 public interface SimpleObject<T extends ValueObject> extends Cloneable, Serializable {
     Object   get(int argIndex);
     Class<?> getFieldClass(int argIndex);
-    byte     getNumberOfFields(); /// This should be static, but stupid Java does not allow it!
+    int     getNumberOfFields(); /// This should be static, but stupid Java does not allow it!
     String[] getTitles();
     T        getValue();
